+++ src/app/Rect.java
@@ -3,15 +3,15 @@ import java.util.Scanner;
 
 public class Rect {
 
-    private float w = 10;
-    private float h = 10;
+    private double w = 10;
+    private double h = 10;
 
-    public Rect(float w, float h) {
+    public Rect(double w, double h) {
         this.w = w;
         this.h = h;
     }
 
-    public Rect(float w) {
+    public Rect(double w) {
         this.w = w;
     }
 
@@ -20,24 +20,24 @@ public class Rect {
     public static Rect create(String text, Scanner in) {
         System.out.println(text);
         System.out.print("Введите длину прямоугольника: ");
-        float w = in.nextFloat();
+        double w = in.nextDouble();
         System.out.print("Введите высоту прямоугольника: ");
-        float h = in.nextFloat();
+        double h = in.nextDouble();
 
         return new Rect(w, h);
     }
 
     public static boolean contains(Rect rect1, Rect rect2) { return rect1.w <= rect2.w && rect1.h <= rect2.h; }
 
-    public float per() { return 2 * this.w + 2 * this.h; }
+    public double per() { return 2 * this.w + 2 * this.h; }
 
-    public float area() { return this.w * this.h; }
+    public double area() { return this.w * this.h; }
 
-    public void setW(float w) { this.w = w; }
+    public void setW(double w) { this.w = w; }
 
-    public void setH(float h) { this.h = h; }
+    public void setH(double h) { this.h = h; }
 
-    public float getW() { return this.w; }
+    public double getW() { return this.w; }
 
-    public float getH() { return this.h; }
+    public double getH() { return this.h; }
 }
